Add country occupation to core logic.
Attack phase logic was missing a key element; namely, the country occupation upon defeat of a defending country. This commit adds PlayerOccupy* events to peril-common, as well as the necessary logic to GameModel and the state machine in order to query army occupation information from the attacking player.

Other changes:
- Change all on**Event methods in StateMachineEventHandler to 'onEvent' ... I finally decided that looks much more sane.
- Minor style fixes (add 'final' to a few classes in the 'battle' package).
- Add new keys to PlayerTurnDataCache.

See PERIL-322