Refactor attack country event scheme.
PlayerAttackCountryRequestEvent is sent from core/server -> client. This event includes a pre-computed multimap of possible attack vectors.

Client responds with PlayerAttackCountryResponseRequestEvent detailing the source and target countries for the attack as well as the number of dice the attacker wishes to roll. The process from there is the same as before, except all of the events that used to be called PlayerAttackCountry**Event are now PlayerAttackCountryResponse**Event.

Client may also respond with a PlayerEndAttackPhaseResponseRequestEvent to signify that they do not wish to submit another attack order. This event has the save request-event type as PlayerAttackCountryResponseRequestEvent, so both will qualify as valid responses when being processed by the server.

Additional changes:
- Add missing @RequiredForNetworkSerialization constructors.
- Fix inconsistent type hierarchies among player response success/denied events (most now simply implement PlayerResponseSuccessEvent or PlayerResponseDeniedEvent).
- Remove PlayerEndAttackPhaseDeniedEvent since it is no longer relevant or necessary.

PERIL-339 #Done