Fix loading screen dropped events bug.
Problem:

On MenuToPlayLoadingScreen, after connecting to a server & joining the
game as a player, there is a significant period of time where the play
map is being loaded. During this time, game-related events can & do
arrive from the server, intended for the play screen. However, since the
play screen is not active yet, they are not received.

Solution:

Register a DeadMessage event bus handler, to catch any events that are
not received by any other handler. Only collect ServerEvent's, ignore
any other message types. After the play screen is made active, in the
method which activated the play screen from the MenuToPlayLoadingScreen,
republish these ServerEvent's in FIFO order, so that the active play
screen will receive them, and in the order they were originally
received.