Modify die roll data structure organization for battle results.
Usage of ImmutableSortedSet <DieFaceValue, DieOutcome> was error prone due to the necessary requirement of allowing identical key values (i.e. multiple dice with the same face value in a single roll). This was replaced with ImmutableSet <DieRoll> where DieRoll is a new type representing a tuple of DieFaceValue and DieOutcome.

Add DieRoll, DieFaceValue, and DieOutcome to KryonetRegistration. This greatly simplifies the logic needed to construct a BattleResultPacket from BattleResult.

Additional changes:
- Remove bad assertion from DefaultBattleModel that was causing inappropriate debug failures.