Change country atlas sprite naming scheme.
Current scheme in *.atlas files for country sprites is, e.g.,
countryNameBlue, where the last capitalized section of the word
represents the country's image state (in other words, the name of the
sprite). The country name and image state must be in camel case, with no
spaces.

This scheme has some limitations & is bug-prone. For example, if an
image state needs to have more than one word, such as "Nuclear Waste",
to visually depict a country after a nuke detonation, it would result in
a name such as countryNameNuclearWaste. The problem is, the code for
parsing these names has no way to know in this case where the country
name ends and where the image state begins, only in cases where the
image state is one word. It is also not too readable.

The new naming scheme is Country Name - Image State, which allows for an
arbitrary number of words to be included in the image state and the
country name. It is more readable, and also even allows for hyphenated
country names or images states, for example: Country-Name - Image State
because the new delimeter is " - ", and hyphenated names do not need
spaces. It also simplifies the country name parsing code, as the country
name is already in its final format.

This is a more flexible & less-bug-prone approach.

Update all country *.atlas files to reflect the new naming scheme.