Make play map code more robust.
Before trying to change a country actor's primary or secondary image
state, check to see if that state exists. If it doesn't, don't try and
change to that image state, just log a warning.

This prevents custom maps that may not implement all possible image
states from crashing the entire client application when trying to change
to a non-existent country image state.