Add observer client support to server.
server:MultiplayerController now keeps a separate mapping of clients -> ObserverPackets. Observers receive all notification and success events that players receive, but are never passed to or recognized by Core.

Clients can join as an observer via ObserverJoinGameRequestEvent. The name must not conflict with an existing player or observer and must be valid according to GameSettings.

There is currently no way for a client to request to be re-assigned from player -> observer, although this could be added in the future. Right now, the more straightforward way of doing this would be to just disconnect and reconnect as an observer.

- Add new type: ObserverPacket (subtype of PersonPacket) with default impl DefaultObserverPacket.
- Add new server communicator, ObserverCommunicator; responsible for sending events to observers.
- Add new mapping class for observers, ClientObserverMapping. This class is basically a simplified version of ClientPlayerMapping and serves as a simple wrapper for the underlying BiMap.
- Add logic to MultiplayerController disconnection event handling for disconnecting observers.
- Add tests to MultiplayerControllerTest for observer join logic.