Implement player win/lose logic.
GameModel now checks the status of each player at the end of a turn phase for whether or not they are still in the game or have won. If a player has been eliminated, the player is removed from PlayerModel and a PlayerLoseGameEvent is sent. If a player has won, a PlayerWinGameEvent and internal EndGameEvent are sent.