Fix event handler ordering bug in MultiplayerController.
Fix heisenbug caused by b0b68f3b. PlayerLeaveGameEvent is now processed by both its handler AND the ServerNotificationEvent handler. This caused arbitrary selection of which handler to run first. The leave game event handler needs to be run first so that the player can be removed from the server before the notification is sent to all players.

- Add Integer.MIN_VALUE priority to ServerNotificationEvent handler so that any other handler for a notificiation event will run before it (it should always be run last since its doing the actual event publishing).