Fix server crash on DeterminePlayerTurnOrderEvent.
Because DeterminePlayerTurnOrderEvent was recently changed in
d688c5c6784388a5fe4fa288c389bcb96c9fc908 to send an ImmutableSortedSet
of turn ordered players instead of just an ImmutableSet, server was
crashing during network serialization of this event since
ImmutableSortedSet is not registered in KryonetRegistration.

Kryonet cannot handle serializing ImmutableSortedSet by default, and the
kryo-serializers library does not contain a serializer for it. Write a
custom serializer, ImmutableSortedSetSerializer, and register it in
KryonetRegistration.

Register PlayerPacket#TURN_ORDER_COMPARATOR, which is used in the
network serialization of DeterminePlayerTurnOrderEvent, in
KryonetRegistration.

Rename KryonetRegistration#EXTERNAL collection of classes to be
registered to #OTHER, as it now contains more than one category of
classes (due to the addition of PlayerPacket#TURN_ORDER_COMPARATOR), and
is no longer accurately named.

Improve documentation in KryonetRegistration regarding the #OTHER
collection of classes to be registered, explaining that it contains
either classes external to the net package or classes internal to the
net package that must be manually registered (such as
PlayerPacket#TURN_ORDER_COMPARATOR).