Fix server crash from network serialization error.
In a recent change, d4c00379d9845cdd9587e84281f967d20fb8c00c, some
events including:

CountryOwnerChangedEvent,
DefaultCountryOwnerChangedEvent,
PlayerOccupyCountryResponseSuccessEvent,
PlayerSelectCountryResponseSuccessEvent

were created and/or refactored to use Guava's Optional class, but that
class was not registered in KryonetRegistration for network
serialization, along with it's implementation classes, Optional#Absent &
Optional#Present, which must be instantiated and retrieved with

This results in the following root error:

com.esotericsoftware.kryo.KryoException:
  java.lang.IllegalArgumentException:
    Class is not registered: com.google.common.base.Absent

Register Optional, Optional#Absent & Optional#Present in
KryonetRegistration. Although the error is only regarding
Optional#Absent, that just happened to be the first point of failure.
All of these three classes need to be registered to solve this crash
completely.