Fix duplicate event bugs on server.
Due to the recent change, d4c00379d9845cdd9587e84281f967d20fb8c00c,
whereby PlayerReinforceCountriesRequestEvent was made to indirectly
extend ServerNotificationEvent, in addition to already extending
PlayerInputRequestEvent, PlayerReinforceCountriesRequestEvent now has
two valid @Handler methods in server::MultiplayerController, one for
PlayerInputRequestEvent that will send out the event to the relevant
player, and the other for ServerNotificationEvent, which will send out
the event to all players (and spectators), resulting in every
PlayerReinforceCountriesRequestEvent being sent out twice from the
server over the network.

The (temporary) solution is to only send the event to all players in the
ServerNotificationEvent @Handler method. To prevent the event from being
sent out by the PlayerInputRequestEvent @Handler method, add an
instanceof check to prevent PlayerInputRequestEvent's which also happen
to be ServerNotificationEvent's, such as
PlayerReinforceCountriesRequestEvent, to be sent out there.