Fix event type doublespeak in core.
The duplicate type/event issue reported in PERIL-384 was being caused by a type concept inconsistency in the refactored event system. The unintended corollary to making events like PlayerReinforceCountriesRequestEvent inherit from PlayerArmiesChangedEvent (or other notification events) is that it itself becomes a notification event.

This is a problem not only because it causes issues in server:MultiplayerController event handling but also because it doesn't make any sense. A ServerRequestEvent is about the farthest thing from a ServerNotificationEvent, and it's a bastardization of the event type system to have an event that inherits both of them. It's basically programming doublespeak ;)

- Remove PlayerArmiesChangedEvent inheritance from PlayerReinforceCountriesRequestEvent.
- Remove new evil (and now thankfully unnecessary) 'instanceof' check from MultiplayerController PlayerInputRequestEvent
  handler.
- Add DefaultPlayerArmiesChangedEvent publication to beginReinforcementPhase() in GameModel.

Additional changes for PERIL-384:
- Add PlayerDeniedEvent as inheriting interface for PlayerResponseDeniedEvent.
- Remove redundant PlayerResponseDeniedEvent handler from server:MultiplayerController.

PERIL-384 #Done