Fix state machine transition bug.
The state transition for reinforcement phase was previously configured to respond to the client *ResponseRequestEvent with and action AND a condition; the condition only passing when country reinforcements and trade-ins had been processed. The action for the transition called the reinforcement processing method in GameModel.

However, this relationship was circular. The state machine will not call a transition action if the condition fails, thus the reinforcements were never processed and the transition never occurred.

This was fixed by returning to checking the verification result as the transition condition. GameModel.isReinforcementComplete() and the corresponding cache entries have been removed.

Note: This means that the trade-in response MUST be sent BEFORE the reinforcement response (which makes sense anyway). Trade-in response events received after the state transition has started will be ignored.

See PERIL-397 #time 20m