Change initial country reinforcement count.
Initial country reinforcement request is now sent with a pre-determined (by game rules) army count for number of reinforcements that can be placed on a country in that turn. For classic mode, this value is 1.

The response consists of only the name of the country to be reinforced. The appropriate number of armies will then be deducted from the player's hand and placed on the country.

See PERIL-371 #time 10m