Create ActivePlayerChangedEvent.
Notify all clients (players & spectators) when the active player has
changed, by publishing a new type of ServerNotificationEvent called
ActivePlayerChangedEvent from GameModel.

The event is published in the following two scenarios:

1) When a PlayerInputRequestEvent is sent to a player, that player
becomes the active player because the server is waiting for an input
response from that player's client. This currently covers two special
game states, manual initial country assignment, and manual initial
reinforcement, or more specifically, upon the publishing of those game
states' respective PLayerInputRequestEvent's:
PlayerClaimCountryRequestEvent &
PlayerReinforceInitialCountryRequestEvent.

2) When a player's turn officially begins upon the publication of
BeginPlayerTurnEvent, which marks the beginning of that player's "turn
phase", which includes reinforcement phase (includes trade-ins), attack
phase, & fortification phase. The player remains active until the end of
the fortification phase, whereby the next player's turn phase begins.

Notifying all clients of when the active player has changed is important
because only one client (the one belonging to the player whose input is
needed) receives a PlayerInputRequestEvent, preventing the other clients
from being able to know (and therefore visually depict) which player is
currently taking action during a game, or which player is being waited
upon by the others for the game to progress.

BeginPlayerTurnEvent took care of scenario #2, but applying
BeginPlayerTurnEvent to scenario #1 would require a massive refactoring
of the entire game logic state machine in core.

Notes:

Note 1: These changes do not consider the defending player during a
battle to be "active"; rather, only the attacker is considered active
because attacking is part of the attacking player's turn phase, not the
defender's. This is actually somewhat arbitrary as a
PlayerDefendCountryRequestEvent (which is a PlayerInputRequestEvent) is
sent out to the defending player, essentially asking that player for
input as well as the attacker. So it could be argued that both the
attacker and defender are "active" during a battle. For now, mainly for
simplicity (and YAGNI), the defender is NOT "active" during a battle.

Note 2: In the special case of when a player wins the game, there is no
ActivePlayerChangedEvent published because there is no point. A player
must already be active (scenario #2 - during their turn phase) in order
to win the game, so the active player does not change in this case.

Note 3: In the special case of when a player loses the game (is
eliminated), there is no ActivePlayerChangedEvent published because
there is no point. An active player cannot lose the game, and players
are eliminated during attack phase only, during which the active player
does not change. If a player loses the game and it results in the game
being over (another player won the game), see note 2.

Miscellaneous:

- Verify publishing of ActivePlayerChangedEvent in GameModelTest for
  manual country assignment by updating existing tests. (No tests
  currently exists for manual initial reinforcement).

- Fix failing tests.

PERIL-697 #Done