Fix message box bugs in client.
Bug 1 - Corrupted message box row assets:

This bug manifests in the play screen quit dialog as corrupted message
text, and is triggered upon leaving and rejoining the play screen. See
PERIL-713.

DefaultMessageBox#refreshAssets method was not refreshing any
MessageBoxRow's assets by calling MessageBoxRow#refreshAssets on the row
instance because the rows are first obtained as raw Actors, which are
never instances of MessageBoxRow's; however, they are required to be in
order to cast them to MessageBoxRow to access
MessageBoxRow#refreshAssets. The solution is simple due to already
having a populated map of Actor's to MessageBoxRow's: obtain the
MessageBoxRow using the map instead of checking it's type or attempting
to cast it.

Bug 2 - Message box memory leak:

DefaultMessageBox caches MessageBoxRow's. However, when old rows are
removed from the MessageBox's table, those rows were not being removed
from the cache. This would lead to the cache continually growing as more
and more messages arrived in the message box (the play screen's chat box
or status box, for example.) The solution is simple: remove the row from
the cache immediately after it is removed in
DefaultMessageBox#limitOldRows.

PERIL-713 #Done