Add request event caching to Core.
InternalCommunicationHandler will now cache PlayerInputRequestEvent types attached to internal request events sent by server::MultiplayerController. This allows Core to easily retrieve the original request event associated with a response and resend it if the response is denied.

NOTE: This commit does NOT implement this thoroughly throughout GameModel. This should be a subsequent task for finishing up PERIL-373