Fix player join game desync bug.
As outlined by PERIL-727, a problem could previously occur where if two players joined simultaneously,
the second player would receive a JoinGameSuccessEvent with the current player list *before* the first
player had been added to it. This meant that the second player never received any information about the
first player joining.

Players will now receive the current player list after joining is complete, i.e. PlayerJoinGameSuccessEvent,
rather than part way through the joining process. This means that the information included in the event
will always be up-to-date with whatever the player status was at the time the success event is constructed.
Because this process is synchronous, one of the two players joining simultaneously will always see the
other in the player set.

A best attempt was made to fix the newly broken code in client. This should be given further review to verify
if anything else needs to be done.

See PERIL-727 #time 30m