Move PlayerColor to common::game.
There are many places in the client that need an actual set of valid
colors to choose from. The workaround (read: hack) was to just whitebox
core and create a list of String colors based on the knowledge of the
contents of PlayerColor in an unavailable module.

This aligns with the larger design intent of moving game-related enum
types out of core and into common::game. Most of them are already moved,
but there are a few that still need to be moved, to be able to access
the full set of valid enum values, even when not connected to the
server.