Refactor attack phase event system.
Attack phase now has two cycling stages: attack vector selection and attack
order submission.

At the beginning of attack phase, a PlayerBeginAttackRequestEvent is sent.
This event contains the multimap of valid attack vectors that the client
needs to make a selection. Once the response is accepted, a
PlayerAttackOrderRequestEvent is sent for the attacker to specify how
many dice he wishes to attack with. After the die count is validated,
a PlayerAttackOrderSuccessEvent is sent, followed by one of three result
events:

PlayerAttackVictoryEvent: Indicates that the attacker has defeated all
remaining defender troops and is now ready to occupy the country. An
occupy request will follow this event.

PlayerAttackDefeatEvent: Indicates that the defender has defeated all
armies able to attack from the source country, and the attacker may
no longer pursue this attack vector. Game state thus proceeds back to
the first attack stage, and a PlayerBeginAttackRequestEvent is sent.

PlayerAttackIndecisiveEvent: Indicates that the battle was indecisive
and neither attacker nor defender have yet triumphed. At this point,
another PlayerAttackOrderRequestEvent will be sent.

The attacking player may always respond to an attack order request
with PlayerRetreatOrderRequestEvent. This will cancel the player's
lock-in on the current attack vector and return attack phase to the
first stage.

Additional changes:

- Miscellaneous improvements to attack phase (and related) events.
- Fix redundant player/country army count change event publication.
- Add min/max die count to PlayerDefendCountryRequestEvent.
- Update BattleModelTest to work with new system.
- Modify client code just enough to build with changes.

PERIL-781 #time 12h #comment "Guessing on time log here..."