Fix stale PlayerPacket bug in GameModel.
Publishing of BeginReinforcementPhaseEvent in GameModel violates
PlayerArmiesChangedEvent interface contract specified in the javadoc
that "All implementations of PlayerArmiesChangedEvent should be
constructed with an updated PlayerPacket (accessible via {@link
<b>after</b> the delta is applied."

BeginReinforcementPhaseEvent#getPlayer#getArmiesInHand is always 0. This
is due to using a stale PlayerPacket instance. Get a fresh PlayerPacket
instance so that the armies in hand represents the amount of armies
after the delta is applied.

PERIL-803 #time 30m