Fix stale CountryPacket bug in GameModel.
Publishing of PlayerReinforceCountrySuccessEvent in GameModel violates
the implicit CountryArmiesChangedEvent interface contract that all
implementations of CountryArmiesChangedEvent should be constructed with
an updated CountryPacket after the delta is applied.

PlayerReinforceCountrySuccessEvent#getCountry#getArmyCount is outdated
when published from GameModel#handlePlayerReinforceCountry. This is due
to using a stale CountryPacket instance. Get a fresh CountryPacket
instance so that the country army count represents the count after the
delta is applied.

Make the contract explicit in CountryArmiesChangedEvent javadoc, which
parallels the PlayerArmiesChangedEvent javadoc interface contract.

PERIL-803 #time 15m