Fix player being denied valid attack vector.
BattleModel inejcted to GameModel by Builder was stale and did not reference
the same instance of PlayMapModel.

Note: This outlines the need for better inversion of control, ideally via an
injection framework. The builder pattern yields too much verbosity and too
many obnoxious bugs like this one. That's the real fix. This fix is a band-aid.

An additional band-aid fix would be to refactor some of the Builder types to
lazily create default instances in the call to Builder#build() rather than in
the constructor. This would help mitigate these kinds of problems; it does not
however remove the need for improved IoC.

PERIL-819 #Done #time 10m