Finish simple AI player.
AI doesn't say much yet, but can play a game all the way through, with
random moves, and lightning speed.

The speed is an issue because it is way beyond the realm of human
comprehension. The only evidence that anything happened is the history
in the status box, and maybe some play map changes.

The client will need to limit the speed of game logic updates.

Note: See PERIL-842, which is causing a state machine crash that seems
to only be triggered with AI. This is currently preventing playing full
games with AI.

Other:

- Minor AI chat improvements.
- Add some event convenience methods to ease AI processing.
- Reduce AI log spam (set some more log levels to WARN).