Implement disconnect/rejoin event handling client-side.
Add DisconnectFromServerEvent (mostly for testing) and ReconnectToServerEvent.

Add disconnect and auto-reconnect logic to client::MultiplayerController.
NOTE: The implementation is not ideal and can be cleaned up. That being said,
it certainly seems to work more naturally in MultiplayerController than in the
play screen.

Add file caching mechanism for active game "sessions" (when successfully joining
a server). Client serializes the data to a file in the settings/ directory. If
the file exists when client is started (presumably due to some kind of crash or
client-side outage), client will show a dialog to the user asking if they want to
reconnect. The serialized data includes the player name, secret id, and server
configuration.

TODO:
- Client needs to properly handle graceful disconnects from the server (quitting)
and game over occurrences. The cache file should be deleted via
GameServerCacheManager::deleteCache to prevent the client from showing the reconnect
dialog the next time it starts.
- Client needs to launch the play loading screen when reconnecting from the main menu.
This will require modifying the loading screen logic to allow for a PlayerJoinGameSuccessEvent
to be constructed with the player secret ID. ReconnectToServerEvent may or may not be necessary
in this case.

Additional changes:
- Update join event toStrings.
- Add confirmation dialog to UI code for showing a yes/no message.

PERIL-389 #Done #time 3h