Consolidate and fix restore/resume events.
:: Core ::
PlayerRestoreGameStateEvent has now been merged into GameResumedEvent,
and GameResumedEvent inherits from new type PlayerRestoreGameStateEvent.
Currently, it is the only necessary implementation of this type, but the
default/abstract implementation should make future expansions of this
functionality easier.

:: Server ::
Note that server only sends a ResumeGameEvent to Core when all players
are bound after a reconnect. Thus, in order to inform reconnected
clients when the game has not yet been resumed, it will now send a new
GameSuspendedEvent *only to the reconnected client.* This is a bit of a
hack on server's part, but a necessary one.

:: Client ::
Move all phase handler activation logic to resume event handler and
remove all references to PlayerRestoreGameStateEvent. Note that the
client still appears to be broken on reconnect; it does not properly
reset the armies/cards in hand and does not display the correct turn
order.

The fix does work in a live client/server game; i.e. the current phase
reported to the client is accurate.

Additional changes:
- Fix resume causing initial game phase to be "restarted" allowing other
  players to prematurely perform their turns before the reconnected
  player. This was caused by the event handlers in
  core::AbstractGamePhaseHandler not being invoked in the right order
  with the state machine, thus allowing the begin/end lifecycle methods to
  be called erroneously when suspending/resuming the game.

PERIL-896 #time 1h #In Progress - Review